+++ inc/filters.php
@@ -9,3 +9,17 @@
  */
 
 defined( 'ABSPATH' ) || exit;
+
+/**
+ * Add allowed tags
+ *
+ * Mainly for Elementor heading widget sanitize function
+ */
+add_filter( 'wp_kses_allowed_html', function( $array ) {
+
+	$array['distorted'] = array();
+	$array['elipse'] = array();
+	$array['underline'] = array();
+
+	return $array;
+}, 44 );